requirements.txt
# Base ----------------------------------------
matplotlib
numpy
opencv-python-headless
Pillow
PyYAML
requests
ultralytics
scipy
torch
torchvision
tqdm
protobuf<4.21.5  # https://github.com/ultralytics/yolov5/issues/8012
pandas
seaborn
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation
streamlit
wget
-e .

